+++ src/transports/sentry_transport_curl.c
@@ -11,7 +11,6 @@
 
 #include <curl/curl.h>
 #include <curl/easy.h>
-#include <stdlib.h>
 #include <string.h>
 
 typedef struct curl_transport_state_s {
@@ -65,6 +64,32 @@ sentry__curl_transport_start(
         CURLcode rv = curl_global_init(CURL_GLOBAL_ALL);
         if (rv != CURLE_OK) {
             SENTRY_WARNF("`curl_global_init` failed with code `%d`", (int)rv);
+            return 1;
+        }
+
+        curl_version_info_data *version_data
+            = curl_version_info(CURLVERSION_NOW);
+
+        if (!version_data) {
+            SENTRY_WARN("Failed to retrieve `curl_version_info`");
+            return 1;
+        }
+
+        sentry_version_t curl_version = {
+            .major = (version_data->version_num >> 16) & 0xff,
+            .minor = (version_data->version_num >> 8) & 0xff,
+            .patch = version_data->version_num & 0xff,
+        };
+
+        if (!sentry__check_min_version(
+                curl_version, (sentry_version_t) { 7, 10, 7 })) {
+            SENTRY_WARNF("`libcurl` is at unsupported version `%u.%u.%u`",
+                curl_version.major, curl_version.minor, curl_version.patch);
+            return 1;
+        }
+
+        if ((version_data->features & CURL_VERSION_ASYNCHDNS) == 0) {
+            SENTRY_WARN("`libcurl` was not compiled with feature `AsynchDNS`");
             return 1;
         }
     }